+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"firebase_auth","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth-5.1.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core-3.2.0/","native_build":true,"dependencies":[]},{"name":"google_sign_in_ios","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/google_sign_in_ios-5.7.6/","shared_darwin_source":true,"native_build":true,"dependencies":[]}],"android":[{"name":"firebase_auth","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth-5.1.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core-3.2.0/","native_build":true,"dependencies":[]},{"name":"google_sign_in_android","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/google_sign_in_android-6.1.28/","native_build":true,"dependencies":[]}],"macos":[{"name":"firebase_auth","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth-5.1.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core-3.2.0/","native_build":true,"dependencies":[]},{"name":"google_sign_in_ios","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/google_sign_in_ios-5.7.6/","shared_darwin_source":true,"native_build":true,"dependencies":[]}],"linux":[],"windows":[{"name":"firebase_auth","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth-5.1.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core-3.2.0/","native_build":true,"dependencies":[]}],"web":[{"name":"firebase_auth_web","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth_web-5.12.4/","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core_web-2.17.3/","dependencies":[]},{"name":"google_sign_in_web","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/google_sign_in_web-0.12.4+1/","dependencies":[]}]},"dependencyGraph":[{"name":"firebase_auth","dependencies":["firebase_auth_web","firebase_core"]},{"name":"firebase_auth_web","dependencies":["firebase_core","firebase_core_web"]},{"name":"firebase_core","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","dependencies":[]},{"name":"google_sign_in","dependencies":["google_sign_in_android","google_sign_in_ios","google_sign_in_web"]},{"name":"google_sign_in_android","dependencies":[]},{"name":"google_sign_in_ios","dependencies":[]},{"name":"google_sign_in_web","dependencies":[]}],"date_created":"2024-07-30 20:52:38.351921","version":"3.22.3"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"firebase_auth","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth-5.1.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core-3.2.0/","native_build":true,"dependencies":[]},{"name":"google_sign_in_ios","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/google_sign_in_ios-5.7.6/","shared_darwin_source":true,"native_build":true,"dependencies":[]}],"android":[{"name":"firebase_auth","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth-5.1.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core-3.2.0/","native_build":true,"dependencies":[]},{"name":"google_sign_in_android","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/google_sign_in_android-6.1.28/","native_build":true,"dependencies":[]}],"macos":[{"name":"firebase_auth","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth-5.1.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core-3.2.0/","native_build":true,"dependencies":[]},{"name":"google_sign_in_ios","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/google_sign_in_ios-5.7.6/","shared_darwin_source":true,"native_build":true,"dependencies":[]}],"linux":[],"windows":[{"name":"firebase_auth","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth-5.1.2/","native_build":true,"dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core-3.2.0/","native_build":true,"dependencies":[]}],"web":[{"name":"firebase_auth_web","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_auth_web-5.12.4/","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/firebase_core_web-2.17.3/","dependencies":[]},{"name":"google_sign_in_web","path":"/Users/qambarabbas/.pub-cache/hosted/pub.dev/google_sign_in_web-0.12.4+1/","dependencies":[]}]},"dependencyGraph":[{"name":"firebase_auth","dependencies":["firebase_auth_web","firebase_core"]},{"name":"firebase_auth_web","dependencies":["firebase_core","firebase_core_web"]},{"name":"firebase_core","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","dependencies":[]},{"name":"google_sign_in","dependencies":["google_sign_in_android","google_sign_in_ios","google_sign_in_web"]},{"name":"google_sign_in_android","dependencies":[]},{"name":"google_sign_in_ios","dependencies":[]},{"name":"google_sign_in_web","dependencies":[]}],"date_created":"2024-07-30 22:18:34.119382","version":"3.22.3"}
